# add proxy_fit_metric_map, speed up simulations by 4x