@Innixma 2023-05-29
Updated so the default is None, meaning the proxy metric is not used unless explicitly set by the user.