@Innixma 2023-05-29
I've updated the code to respond to comments and added a TODO for a unit test.

I'm merging to simplify the amount of in-progress pull requests.